+++ docs/examples/s2i/configs/redis.yaml
@@ -1,0 +1,7 @@
+name: redis
+controller: deploymentconfig
+containers:
+- image: redis
+services:
+- portMappings:
+  - "6379"
